+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHXCA9PJ6ZP7P4CMBXJVVYG0"
+        "01KHXEZ291J66ZGSWANXMM3NHZ"
     )
     r"""The unique identifier of the data source"""