+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHXCA9PFF7CH2VZ1TYGNP3YV"
+        "01KHXEZ28XAR5VZGJKY2S107R1"
     )
     r"""The unique identifier of the data source"""