+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHXCA9PW5PDGGJ205J5V7D9Y"
+        "01KHXEZ295KFM40CZF71NQAR1R"
     )
     r"""The unique identifier of the data source"""